+++ ris_texture_toolkit_c/cmake/VcpkgDependencies.cmake
@@ -1,0 +1,18 @@
+
+find_package(ktx CONFIG REQUIRED)
+find_package(spdlog CONFIG REQUIRED)
+find_package(Vulkan REQUIRED)
+
+SET(PACKAGES_LIBRARIES
+        ${PACKAGES}
+        ${SLANG_LIBRARY}
+        KTX::ktx
+        spdlog::spdlog
+        Vulkan::Vulkan
+)
+
+SET(PACKAGES_INCLUDE
+        ${PACKAGES_INCLUDE}
+        ${SLANG_INCLUDE_DIR}
+)
+